service-intelligence/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
grpcio==1.63.0
protobuf==5.26.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
aiosqlite==0.20.0
greenlet==3.0.3
numpy>=1.26.0
scipy>=1.12.0
httpx>=0.27
